requirements.txt
fastmcp
langchain
langchain-core
langchain-community
langchain-ollama
httpx
nest-asyncio